% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bulls}
\alias{bulls}
\title{The Chicago Bulls player stats}
\format{
## `bulls`
A data frame with 108 rows and 7 columns:
\describe{
  \item{...1}{player number on team}
  \item{player}{player name}
  \item{year}{the year of the season}
  \item{RVORP}{regular season adjusted VORP}
  \item{PVORP}{post season adjusted VORP}
  \item{win}{whether they won the championship that season}
  \item{DVORP}{the difference in playoff and regular season VORP}
}

## `bulls`
A data frame with 108 rows and y columns:
\describe{
  \item{...1}{player number on team}
  \item{player}{player name}
  \item{year}{the year of the season}
  \item{RVORP}{regular season adjusted VORP}
  \item{PVORP}{post season adjusted VORP}
  \item{win}{whether they won the championship that season}
  \item{DVORP}{the difference in playoff and regular season VORP}
}
}
\source{
by using the functions in this package

by using the functions in this package
}
\usage{
bulls

bulls
}
\description{
The Chicago Bulls player stats

The Chicago Bulls player stats
}
\keyword{datasets}
