% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{caveliers}
\alias{caveliers}
\title{The Cleveland Caveliers player stats}
\format{
## `caveliers`
A data frame with 67 rows and 7 columns:
\describe{
  \item{...1}{player number on team}
  \item{player}{player name}
  \item{year}{the year of the season}
  \item{RVORP}{regular season adjusted VORP}
  \item{PVORP}{post season adjusted VORP}
  \item{win}{whether they won the championship that season}
  \item{DVORP}{the difference in playoff and regular season VORP}
}
}
\source{
by using the functions in this package
}
\usage{
caveliers
}
\description{
The Cleveland Caveliers player stats
}
\keyword{datasets}
