% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celtics}
\alias{celtics}
\title{The Boston Celtics player stats}
\format{
## `celtics`
A data frame with 131 rows and 7 columns:
\describe{
  \item{...1}{player number on team}
  \item{player}{player name}
  \item{year}{the year of the season}
  \item{RVORP}{regular season adjusted VORP}
  \item{PVORP}{post season adjusted VORP}
  \item{win}{whether they won the championship that season}
  \item{DVORP}{the difference in playoff and regular season VORP}
}
}
\source{
by using the functions in this package
}
\usage{
celtics
}
\description{
The Boston Celtics player stats
}
\keyword{datasets}
