% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_wrapper.R
\name{cleaning}
\alias{cleaning}
\title{Title}
\usage{
cleaning(decade, team, list)
}
\arguments{
\item{decade}{the years wanted}

\item{team}{the team wanted}

\item{list}{an empty list to fill}
}
\value{
the list of players from specific team wanted
}
\description{
Title
}
\examples{
data("playoff_games")
data("regular_season")
data("names")

names(playoff_games) <- names
names(regular_season) <- names

three_teams_80s <- filter_teams1(1980,1990,regular_season,'BOS','DET','LAL')

output <- list()
boston <- cleaning(three_teams_80s, 'BOS', output)
}
