% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_teams2_wrapper.R
\name{filter_teams2}
\alias{filter_teams2}
\title{Title}
\usage{
filter_teams2(first, last, part, team1, team2, team3, team4)
}
\arguments{
\item{first}{}

\item{team4}{}
}
\value{
the adjusted vorp of players
}
\description{
Title
}
