% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heat}
\alias{heat}
\title{The Miami Heat player stats}
\format{
## `heat`
A data frame with 90 rows and 7 columns:
\describe{
  \item{...1}{player number on team}
  \item{player}{player name}
  \item{year}{the year of the season}
  \item{RVORP}{regular season adjusted VORP}
  \item{PVORP}{post season adjusted VORP}
  \item{win}{whether they won the championship that season}
  \item{DVORP}{the difference in playoff and regular season VORP}
}
}
\source{
by using the functions in this package
}
\usage{
heat
}
\description{
The Miami Heat player stats
}
\keyword{datasets}
