% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names}
\alias{names}
\title{The names of the dataframe's columns}
\format{
## `names`
A vector of names
}
\source{
custom picked column names
}
\usage{
names
}
\description{
The names of the dataframe's columns
}
\keyword{datasets}
