% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_wrapper.R
\name{player}
\alias{player}
\title{Title}
\usage{
player(play, team, titles = "Player Performance")
}
\arguments{
\item{play}{the desired player}

\item{team}{the team wanted}

\item{titles}{the title of the graph}
}
\value{
a plot comparing a player's regular season to post season
}
\description{
Title
}
\examples{
data("lakers")

player('abdulka01',lakers)
}
