% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{playoff_games}
\alias{playoff_games}
\title{The list of all recorded post season NBA games}
\format{
## `playoff_games`
A data frame with 5906 rows and 11 columns:
\describe{
  \item{Visitor.Neutral}{the visiting team}
  \item{Visitor.Points}{visiting team's points}
  \item{Home.Neutral}{the home team}
  \item{Home.Points}{home team's points}
  \item{id}{game id}
  \item{year}{year the game was played in}
  \item{win}{who won the game}
  \item{team}{the team who won}
  \item{season_id}{which season it was in}
  \item{margin}{margin of points}
  \item{playoff}{was the game in the playoffs}
}
}
\source{
by using the functions in this package
}
\usage{
playoff_games
}
\description{
The list of all recorded post season NBA games
}
\keyword{datasets}
