% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_wrapper.R
\name{post}
\alias{post}
\title{Title}
\usage{
post(
  player1,
  player2,
  team1,
  team2,
  player3 = NULL,
  team3 = NULL,
  player4 = NULL,
  team4 = NULL,
  mine = NULL,
  mine2 = NULL
)
}
\arguments{
\item{player1}{the first player wanting to be compared}

\item{player2}{the second player wanting to be compared}

\item{team1}{the first player's team}

\item{team2}{the second player's team}

\item{player3}{the third player wanting to be compared}

\item{team3}{the third player's team}

\item{player4}{the fourth player wanting to be compared}

\item{team4}{the fourth player's team}

\item{mine}{the first player's name}

\item{mine2}{the second player's name}
}
\value{
a line graph comparing two player's post seasons
}
\description{
Title
}
\examples{
data("jazz")
data("bulls")

post('jordami01','malonka01',bulls,jazz, mine = 'Michael Jordan',mine2 = 'Karl Malone')
}
