% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reg.bos.80}
\alias{reg.bos.80}
\title{The Golden State Warriors player stats}
\format{
## `reg.bos.80`
A list containing multiple years of the Boston Celtics stats
\describe{
  \item{list}{list of the stats}
}
}
\source{
by using the functions in this package
}
\usage{
reg.bos.80
}
\description{
The Golden State Warriors player stats
}
\keyword{datasets}
