% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_wrapper.R
\name{reg}
\alias{reg}
\title{Title}
\usage{
reg(
  player1,
  player2,
  team1,
  team2,
  player3 = NULL,
  team3 = NULL,
  player4 = NULL,
  team4 = NULL
)
}
\arguments{
\item{player1}{the first player wanting to be compared}

\item{player2}{the second player wanting to be compared}

\item{team1}{the first player's team}

\item{team2}{the second player's team}

\item{player3}{the third player wanting to be compared}

\item{team3}{the third player's team}

\item{player4}{the fourth player wanting to be compared}

\item{team4}{the fourth player's team}
}
\value{
a line graph comparing two player's regular seasons
}
\description{
Title
}
\examples{
data("jazz")
data("warriors")

reg('curryst01','malonka01',warriors,jazz)
}
