% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regular_season}
\alias{regular_season}
\title{The list of all recorded regular season NBA games}
\format{
## `regular_season`
A data frame with 86534 rows and 11 columns:
\describe{
  \item{Visitor.Neutral}{the visiting team}
  \item{Visitor.Points}{visiting team's points}
  \item{Home.Neutral}{the home team}
  \item{Home.Points}{home team's points}
  \item{id}{game id}
  \item{year}{year the game was played in}
  \item{win}{who won the game}
  \item{team}{the team who won}
  \item{season_id}{which season it was in}
  \item{margin}{margin of points}
  \item{playoff}{was the game in the playoffs}
}
}
\source{
by using the functions in this package
}
\usage{
regular_season
}
\description{
The list of all recorded regular season NBA games
}
\keyword{datasets}
