% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatt_wrapper.R
\name{scatt}
\alias{scatt}
\title{Title}
\usage{
scatt(data, years = NULL, mine = NULL)
}
\arguments{
\item{data}{the dataframe of the team interested}

\item{years}{the years that is wanted to see}

\item{mine}{the plot's title}
}
\value{
scatterplot of a teams perfomance comparing the adjusted vorp of the regular season to the post season
}
\description{
Title
}
\examples{
data("heat")

scatt(heat)
}
