% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spurs}
\alias{spurs}
\title{The San Antonio Spurs player stats}
\format{
## `spurs`
A data frame with 254 rows and 7 columns:
\describe{
  \item{...1}{player number on team}
  \item{player}{player name}
  \item{year}{the year of the season}
  \item{RVORP}{regular season adjusted VORP}
  \item{PVORP}{post season adjusted VORP}
  \item{win}{whether they won the championship that season}
  \item{DVORP}{the difference in playoff and regular season VORP}
}
}
\source{
by using the functions in this package
}
\usage{
spurs
}
\description{
The San Antonio Spurs player stats
}
\keyword{datasets}
