% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorp_wrapper.R
\name{vorp}
\alias{vorp}
\title{Title}
\usage{
vorp(df)
}
\arguments{
\item{df}{the first year desired}
}
\value{
a large csv of the players performance over the decade
}
\description{
Title
}
\examples{
data("reg_bos_80")

test <- reg.bos.80[[1]][1,]

vorp(test)
}
