% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_wrapper.R
\name{cleaning}
\alias{cleaning}
\title{Title}
\usage{
cleaning(decade, team, list)
}
\arguments{
\item{decade}{the years wanted}

\item{team}{the team wanted}

\item{list}{}
}
\value{
the list of players from the team
}
\description{
Title
}
