% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_teams1_wrapper.R
\name{filter_teams1}
\alias{filter_teams1}
\title{Title}
\usage{
filter_teams1(first, last, part, team1, team2, team3)
}
\arguments{
\item{first}{}

\item{team3}{}
}
\value{
the adjusted vorp of players
}
\description{
Title
}
