% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_wrapper.R
\name{player}
\alias{player}
\title{Title}
\usage{
player(play, team, titles = "Player Performance")
}
\arguments{
\item{team}{the team wanted}

\item{decade}{the years wanted}

\item{list}{}
}
\value{
the list of players from the team
}
\description{
Title
}
