% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_wrapper.R
\name{post}
\alias{post}
\title{Title}
\usage{
post(
  player1,
  player2,
  team1,
  team2,
  player3 = NULL,
  team3 = NULL,
  player4 = NULL,
  team4 = NULL,
  mine = NULL,
  mine2 = NULL
)
}
\arguments{
\item{decade}{the years wanted}

\item{team}{the team wanted}

\item{list}{}
}
\value{
the list of players from the team
}
\description{
Title
}
