% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatt_wrapper.R
\name{scatt}
\alias{scatt}
\title{Title}
\usage{
scatt(data, years = NULL, mine = NULL)
}
\arguments{
\item{decade}{the years wanted}

\item{team}{the team wanted}

\item{list}{}
}
\value{
the list of players from the team
}
\description{
Title
}
